Move bmct and all_properties to jbmc/
These aren't used by CBMC anymore, but only
for the symex-driven lazy-loading mode of JBMC.
