+++ jbmc/src/jbmc/all_properties.cpp
