+++ jbmc/src/jbmc/jbmc_parse_options.cpp
@@ -61,7 +61,7 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include <langapi/mode.h>
 
-#include <cbmc/bmc.h> // will go away
+#include "bmc.h" // will go away
 
 #include <java_bytecode/convert_java_nondet.h>
 #include <java_bytecode/java_bytecode_language.h>
